- clean up imports on test_routers module - move get_merchant_cls to a utils module. We'll need it when implementing our SaaSAPIView
